Refactor home page to use client hero component
