+++ lib/courseData.js
@@ -1,0 +1,124 @@
+export const benefits = [
+  'Разработка КД и 3D‑моделей по существующим деталям',
+  'Оперативный ремонт и изготовление запчастей на АТ',
+  'Снижение издержек с помощью собственной оснастки',
+]
+
+export const audience = [
+  'Инженеры‑конструкторы и технологи',
+  'Операторы ЧПУ',
+  'Студенты техвузов, молодые учёные, преподаватели',
+  'Промышленные дизайнеры',
+]
+
+export const modules = [
+  {
+    day: '01 (Пн)',
+    blocks: [
+      {
+        title:
+          'Лекция: Реверсивный инжиниринг и аддитивные технологии в производстве. ОТ и ТБ. Кейсы РГСУ (Hi‑Tech, Ростех, Северсталь, СИБУР, ЕВРАЗ) ',
+        hours: '2 ч (1 лк + 0,5 ем + 0,5 кт)',
+        control: 'Устный опрос',
+      },
+      {
+        title: 'Мастер‑класс #1: CAD/CAM‑моделирование мастер‑моделей и метаформ (T‑FLEX CAD)',
+        hours: '—',
+        control: 'Зачёт',
+      },
+      {
+        title: 'Прак‑работа #1: Выбор средства оцифровки и оснастки. Калибровка сканера (RangeVision Spectrum)',
+        hours: '2 ч',
+        control: '—',
+      },
+      { title: 'Прак‑работа #2: 3D‑сканирование на стационарном сканере', hours: '4 ч', control: '—' },
+    ],
+  },
+  {
+    day: '02 (Вт)',
+    blocks: [
+      { title: 'Прак‑работа #3: Реверс в Geomagic Design X (базовые функции)', hours: '4 ч', control: '—' },
+      {
+        title: 'Мастер‑класс #2: Основы 3D‑печати (FDM, PICASO 3D). От индустриального партнёра',
+        hours: '2 ч (0,5 лк + 1 пр + 0,5 кт)',
+        control: '—',
+      },
+      { title: 'Прак‑работа #4: Подготовка моделей к FDM/DLP/SLA. Запуск печати', hours: '2 ч', control: '—' },
+    ],
+  },
+  {
+    day: '03 (Ср)',
+    blocks: [
+      {
+        title: 'Мастер‑класс #3: 3D‑сканирование ручным оптическим сканером (Artec Eva)',
+        hours: '2 ч (0,5 лк + 1 пр + 0,5 кт)',
+        control: 'Устный опрос',
+      },
+      { title: 'Прак‑работа #5: Geomagic Design X (продвинутые функции)', hours: '6 ч (1 лк + 5 пр)', control: 'Зачёт' },
+    ],
+  },
+  {
+    day: '04 (Чт)',
+    blocks: [
+      {
+        title: 'Мастер‑класс #4: Основы DLP/SLA‑печати. От индустриального партнёра',
+        hours: '2 ч (0,5 лк + 1 пр + 0,5 кт)',
+        control: 'Устный опрос',
+      },
+      { title: 'Прак‑работа #6: Подготовка к FDM/DLP/SLA. Запуск печати', hours: '2 ч (0,5 лк + 1 пр + 0,5 кт)', control: 'Зачёт' },
+      { title: 'Отработка навыков (сканирование/реверс/печать) — свободный формат', hours: '4 ч', control: '—' },
+    ],
+  },
+  {
+    day: '05 (Пт)',
+    blocks: [
+      {
+        title:
+          'Экзамен (ДЭ): задание в формате International High‑Tech Competition (компетенция «Реверсивный инжиниринг»)',
+        hours: '16 ч (2 лк + 12 пр + 2 кт)',
+        control: 'Экзамен (ДЭ)',
+      },
+    ],
+  },
+  { day: '06 (Сб)', blocks: [] },
+]
+
+export const lead = {
+  price: '68\u202f000 ₽',
+  duration: '48 часов (1 неделя, пн‑сб 09:00—18:00)',
+  seats: '6–12 человек в группе',
+  venue: 'Москва, ул. Вильгельма Пика, д. 4, к. 8 (технопарк РГСУ)',
+}
+
+export const mentors = [
+  {
+    name: 'Евгений Сафонов',
+    role: 'Руководитель лаборатории, эксперт по реверсивному инжинирингу',
+    bio: '15 лет в производстве, руководил внедрением аддитивных технологий у промышленного партнёра Ростех.',
+  },
+  {
+    name: 'Марина Боровик',
+    role: 'Ведущий инженер‑технолог',
+    bio: 'Специализируется на интеграции 3D‑сканирования в потоковые процессы, автор учебно‑методических материалов РГСУ.',
+  },
+  {
+    name: 'Дмитрий Казаков',
+    role: 'Инженер‑программист ЧПУ',
+    bio: 'Эксперт по CAD/CAM и T‑FLEX, сертифицированный тренер PICASO 3D.',
+  },
+]
+
+export const faq = [
+  {
+    q: 'Нужен ли свой ноутбук?',
+    a: 'Желательно: c Windows 10/11, 16 ГБ ОЗУ и поддержкой USB 3.0. На площадке доступны рабочие станции, но личное устройство ускорит работу.',
+  },
+  {
+    q: 'Можно ли пройти курс без опыта в CAD?',
+    a: 'Да. Мы стартуем с базовых инструментов и даём дополнительные сессии по интерфейсам программ. Преподаватели помогают на практике.',
+  },
+  {
+    q: 'Вы выдаёте документы?',
+    a: 'По итогу интенсивного курса участники получают удостоверение о повышении квалификации установленного образца РГСУ.',
+  },
+]
